Move codo from main
